Update.
1998-12-16  Andreas Schwab  <schwab@issan.cs.uni-dortmund.de>

	* sysdeps/posix/tempname.c (__path_search): Add casts for format
	string.

1998-12-16  Andreas Schwab  <schwab@issan.cs.uni-dortmund.de>

	* sunrpc/rpc/xdr.h (XDR_DESTROY, xdr_destroy): Avoid dangling else
	problem.
	(IXDR_PUT_INT32): Fully parenthesize parameters.
	(IXDR_GET_LONG, IXDR_PUT_LONG): Mark GCC extension.  Fully
	parenthesize parameter.

+++ sysdeps/posix/tempname.c
@@ -95,7 +95,7 @@ __path_search (char *tmpl, size_t tmpl_len, const char *dir, const char *pfx,
       return -1;
     }
 
-  sprintf (tmpl, "%.*s/%.*sXXXXXX", dlen, dir, plen, pfx);
+  sprintf (tmpl, "%.*s/%.*sXXXXXX", (int) dlen, dir, (int) plen, pfx);
   return 0;
 }
